Make RabbitMQMessageScheme more flexible.
Allow it to be able to be wrapped around any
existing scheme without having to change any code.
You can supply a field name for both properties and 
envelope and access them at the first bolt that
talks to the spout.

If you need information specific to RabbitMQ, you
need to grab it there and thread through the rest
of your topology. This means, that is you were using,
say SimpleJsonScheme, you wouldn't need to
change any code that was looking for the field name:

"object".

Field ordering also renames unchanged with 
envelope and properties being added at the end.