Allow dynamic determination of exchange, routing key, message headers, etc based on incoming tuple (it only allow extraction of message body till now).
This can be accomplished by providing a TupleToMessage implementation to RabbitMQBolt. If all of these fields are fixed and do not change for the message that go through this topology, then you can
use the simplified TupleToMessageNonDynamic to provide fixed values for these fields at topology deployment time via storm configuration.